GradeAntPlus_Prototype/requirements.txt
google-adk
google-genai
google-generativeai
httpx
gradio
pathlib